Assert test name and directory name match
We have a couple of examples of test names and directory names being
slightly different. This runner can show the (incorrect) test name,
misleading the user about where the issue is.

So just fail early and outright if we run into a test with this issue.
This will force people to fix the tests.