Replace `net5.0` TargetFramework in csproj too
When running an xunit project, we patch the TargetFramework in the
csproj file to make sure it targets the runtime we are trying to test.
That's currently implemented for TargetFramework values matching the
regex `netcoreapp\d\.\d`.  Extend the regex to match `net\d\.\d` as
well.

This fixes the first issue caught in
https://github.com/redhat-developer/dotnet-regular-tests/pull/131