Handle NuGet.Config and ProdCon.txt missing
That's not really likely to happen, unless the SSL/HTTP stack itself is
broken. But if that's the case, the runner should try and continue
without the extra configuration.

It's possible to run into this if OpenSSL 3.0 is the system OpenSSL
library and .NET (using OpenSSL 1.1) has some problems using the OpenSSL
3.0 configuration at /etc/ssl/openssl.conf.

Also use Exception.ToString() everywhere to see the full exception stack
traces.