Switch back to portable application and .NET Core 3.1
This change breaks the release format. Now the release is a tarball
containing a single directory `turkey` containing `turkey/Turkey.dll`
which is the main dll of the application.

To support multiple .NET releases, platforms, architectures and even
operating system differences more easily, switch to portable builds with
roll-forward enabled. That allows the system .NET (which presuably works
on the combination of OS, architecture, system libraries, etc) to take
care of running the test suite itself.

That requires switching the application back to .NET Core 3.1 as the
TargetFramework. Enabling RollForward means the application should
continue to work fine on newer releases. This chnage enables running
unit tests to help confirm that.