Fix identifying Mono vs CoreCLR runtimes
We currently use `mono-gc.h` as signal that the runtime is a mono-based
runtime. Unfortuatnely, that file is only included with 7.0. This makes
us mis-identify Mono-based 6.0 runtimes as CoreCLR-based instead.

Instead, use libcoreclrtraceptprovider.so which is available
CoreCLR-based platforms (eg, x86_64) but not Mono-based ones. This file
exists on 6.0 and later:

    $ find /usr/lib64/dotnet -iname libcoreclrtraceptprovider.so
    /usr/lib64/dotnet/shared/Microsoft.NETCore.App/8.0.0-preview.5.23280.8/libcoreclrtraceptprovider.so
    /usr/lib64/dotnet/shared/Microsoft.NETCore.App/7.0.7/libcoreclrtraceptprovider.so
    /usr/lib64/dotnet/shared/Microsoft.NETCore.App/6.0.18/libcoreclrtraceptprovider.so